+++ framework/subsystems/opimd/pimb_sqlite_notes.py
@@ -157,11 +157,14 @@ class SQLiteNotesBackend(Backend):
         for (field, value) in entry_data:
             if field=='_backend_entry_id':
                 entryId=value
+        deleted = []
         for (field, value) in entry_data:
             if field in reqfields:
                 cur.execute('UPDATE notes SET '+field+'=? WHERE id=?',(value,entryId))
             elif not field.startswith('_'):
-                cur.execute('DELETE FROM note_values WHERE noteId=? AND field=?',(entryId,field))
+                if not field in deleted:
+                    cur.execute('DELETE FROM note_values WHERE noteId=? AND field=?',(entryId,field))
+                    deleted.append(field)
                 if isinstance(value, Array) or isinstance(value, list):
                     for val in value:
                         cur.execute('INSERT INTO note_values (field,value,noteId) VALUES (?,?,?)',(field,val,entryId))
